Moved slicing under dataflowengineoss
