+++ build.sbt
@@ -10,7 +10,6 @@ lazy val console           = Projects.console
 lazy val dataflowengineoss = Projects.dataflowengineoss
 lazy val macros            = Projects.macros
 lazy val semanticcpg       = Projects.semanticcpg
-lazy val slicing           = Projects.slicing
 lazy val benchmarks        = Projects.benchmarks
 lazy val c2cpg             = Projects.c2cpg
 lazy val ghidra2cpg        = Projects.ghidra2cpg
@@ -30,7 +29,6 @@ lazy val aggregatedProjects: Seq[ProjectReference] = Seq(
   dataflowengineoss,
   macros,
   semanticcpg,
-  slicing,
   c2cpg,
   x2cpg,
   pysrc2cpg,
