+++ dataflowengineoss/src/main/scala/io/joern/dataflowengineoss/slicing/package.scala
@@ -1,8 +1,7 @@
-package io.joern
+package io.joern.dataflowengineoss
 
 import better.files.File
-import io.circe.{Decoder, Encoder, HCursor, Json}
-import io.joern.slicing.SliceMode.SliceModes
+import io.joern.dataflowengineoss.slicing.SliceMode.SliceModes
 import io.shiftleft.codepropertygraph.generated.PropertyNames
 import io.shiftleft.codepropertygraph.generated.nodes._
