+++ project/Projects.scala
@@ -9,7 +9,6 @@ object Projects {
   lazy val dataflowengineoss = project.in(file("dataflowengineoss"))
   lazy val macros            = project.in(file("macros"))
   lazy val semanticcpg       = project.in(file("semanticcpg"))
-  lazy val slicing           = project.in(file("slicing"))
   lazy val benchmarks        = project.in(file("benchmarks"))
 
   lazy val c2cpg       = project.in(frontendsRoot / "c2cpg")
