--- slicing/README.md
@@ -1,4 +1,0 @@
-# joern-slicing
-
-This module contains logic for extracting subsets of the CPG and serializing them for various downstream tasks. To
-use this from the command line, use the `joern-slice` executable.
